+++ stroom-app/src/test/java/stroom/index/TestIndexShardWriterImpl.java
@@ -17,34 +17,47 @@
 
 package stroom.index;
 
+import stroom.datasource.api.v2.AnalyzerType;
+import stroom.datasource.api.v2.FieldType;
 import stroom.docref.DocRef;
+import stroom.entity.shared.ExpressionCriteria;
 import stroom.index.impl.IndexDocument;
-import stroom.index.impl.IndexShardKeyUtil;
 import stroom.index.impl.IndexShardManager;
 import stroom.index.impl.IndexShardManager.IndexShardAction;
 import stroom.index.impl.IndexShardService;
 import stroom.index.impl.IndexShardWriter;
 import stroom.index.impl.IndexShardWriterCache;
 import stroom.index.impl.IndexStore;
+import stroom.index.impl.IndexVolumeDao;
 import stroom.index.impl.Indexer;
 import stroom.index.shared.FindIndexShardCriteria;
 import stroom.index.shared.IndexException;
 import stroom.index.shared.IndexShard;
 import stroom.index.shared.IndexShard.IndexShardStatus;
 import stroom.index.shared.IndexShardKey;
+import stroom.index.shared.IndexVolume;
 import stroom.index.shared.LuceneIndexDoc;
 import stroom.index.shared.LuceneIndexField;
+import stroom.node.api.NodeInfo;
+import stroom.query.language.functions.ValInteger;
 import stroom.query.language.functions.ValString;
 import stroom.search.extraction.FieldValue;
 import stroom.test.AbstractCoreIntegrationTest;
 import stroom.test.CommonTestControl;
 import stroom.test.CommonTestScenarioCreator;
+import stroom.util.shared.ResultPage;
+import stroom.util.shared.Selection;
 
 import jakarta.inject.Inject;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.attribute.PosixFilePermission;
+import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -65,6 +78,10 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
     private CommonTestControl commonTestControl;
     @Inject
     private IndexStore indexStore;
+    @Inject
+    private IndexVolumeDao indexVolumeDao;
+    @Inject
+    private NodeInfo nodeInfo;
 
     @BeforeEach
     void onBefore() {
@@ -72,7 +89,7 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
     }
 
     @Test
-    void testSingle() throws IOException {
+    void testSingle() {
         assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isZero();
 
         // Do some work.
@@ -82,10 +99,11 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
         // Create an index
         final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010a");
         final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
-        final IndexShardKey indexShardKey1 = IndexShardKeyUtil.createTestKey(index1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+        final IndexShard indexShard = indexShardService.createIndexShard(indexShardKey1, nodeInfo.getThisNodeName());
 
         // Create a writer in the pool
-        final IndexShardWriter writer1 = indexShardWriterCache.getWriterByShardKey(indexShardKey1);
+        final IndexShardWriter writer1 = indexShardWriterCache.getWriter(indexShard.getId());
 
         // Assert that there is 1 writer in the pool.
         assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isEqualTo(1);
@@ -119,7 +137,7 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
     }
 
     @Test
-    void testSimple() throws IOException {
+    void testSimple() {
         assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isZero();
 
         // Do some work.
@@ -128,15 +146,17 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
 
         final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010");
         final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
-        final IndexShardKey indexShardKey1 = IndexShardKeyUtil.createTestKey(index1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+        final IndexShard indexShard1 = indexShardService.createIndexShard(indexShardKey1, nodeInfo.getThisNodeName());
 
         final DocRef indexRef2 = commonTestScenarioCreator.createIndex("TEST_2011");
         final LuceneIndexDoc index2 = indexStore.readDocument(indexRef2);
-        final IndexShardKey indexShardKey2 = IndexShardKeyUtil.createTestKey(index2);
+        final IndexShardKey indexShardKey2 = IndexShardKey.createKey(index2);
+        final IndexShard indexShard2 = indexShardService.createIndexShard(indexShardKey2, nodeInfo.getThisNodeName());
 
         // Create 2 writers in the pool.
-        final IndexShardWriter writer1 = indexShardWriterCache.getWriterByShardKey(indexShardKey1);
-        final IndexShardWriter writer2 = indexShardWriterCache.getWriterByShardKey(indexShardKey2);
+        final IndexShardWriter writer1 = indexShardWriterCache.getWriter(indexShard1.getId());
+        final IndexShardWriter writer2 = indexShardWriterCache.getWriter(indexShard2.getId());
 
         // Assert that there are 2 writers in the pool.
         assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isEqualTo(2);
@@ -188,6 +208,168 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
         assertThat(compareStatus(IndexShardStatus.OPEN, writer2.getIndexShardId())).isFalse();
     }
 
+    @Test
+    void testSimpleRoll() {
+        // Do some work.
+        final IndexDocument document = new IndexDocument();
+        document.add(new FieldValue(LuceneIndexField.createField("test"), ValString.create("test")));
+
+        final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010",
+                commonTestScenarioCreator.createIndexFields(),
+                10);
+        final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+
+        for (int i = 0; i < 10; i++) {
+            indexer.addDocument(indexShardKey1, document);
+        }
+
+        final ResultPage<IndexShard> indexShardResultPage = indexShardService.find(FindIndexShardCriteria.matchAll());
+        assertThat(indexShardResultPage.size()).isOne();
+        final IndexShard indexShard1 = indexShardResultPage.getFirst();
+        final IndexShardWriter writer1 = indexShardWriterCache.getWriter(indexShard1.getId());
+
+        // Make sure the writer is full.
+        assertThatThrownBy(() -> writer1.addDocument(document)).isInstanceOf(IndexException.class);
+
+        // Make sure the writer is still open.
+        assertThat(compareStatus(IndexShardStatus.OPEN, writer1.getIndexShardId())).isTrue();
+
+        // Now push the writer over the edge so we get a new writer.
+        indexer.addDocument(indexShardKey1, document);
+
+        // Get the new writer.
+        final IndexShardWriter writer2 = indexShardWriterCache.getWriter(indexShard1.getId());
+
+        // Make sure the writers are not the same.
+        assertThat(writer2).isNotEqualTo(writer1);
+
+        for (int i = 1; i < 10; i++) {
+            assertThat(indexShardWriterCache.getWriter(indexShard1.getId())).isEqualTo(writer2);
+            indexer.addDocument(indexShardKey1, document);
+        }
+
+        // Make sure the writer is full.
+        assertThatThrownBy(() -> writer2.addDocument(document)).isInstanceOf(IndexException.class);
+
+        // Make sure the writer is still open.
+        assertThat(compareStatus(IndexShardStatus.OPEN, writer2.getIndexShardId())).isTrue();
+    }
+
+    @Test
+    void testFileSystemError() throws IOException {
+        assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isZero();
+
+        // Make index volume paths that we cannot write to.
+        final Path tempDir = Files.createTempDirectory("stroom");
+        Files.setPosixFilePermissions(tempDir, Set.of(PosixFilePermission.OWNER_READ));
+        final ResultPage<IndexVolume> resultPage = indexVolumeDao.find(new ExpressionCriteria());
+        resultPage.forEach(indexVolume -> {
+            indexVolume.setPath(tempDir + indexVolume.getPath());
+            indexVolumeDao.update(indexVolume);
+        });
+
+        // Do some work.
+        final IndexDocument document = new IndexDocument();
+        document.add(new FieldValue(LuceneIndexField.createField("test"), ValString.create("test")));
+
+        final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010",
+                commonTestScenarioCreator.createIndexFields(),
+                10);
+        final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+
+        assertThatThrownBy(() -> indexer.addDocument(indexShardKey1, document))
+                .isInstanceOf(RuntimeException.class)
+                .hasMessageContaining("Index volume path not found");
+    }
+
+    @Test
+    void testChangeFieldType() {
+        assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isZero();
+
+        final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010",
+                commonTestScenarioCreator.createIndexFields(),
+                10);
+        final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+
+        final Selection<IndexShardStatus> statusSelection = Selection.selectNone();
+        statusSelection.addAll(Set.of(IndexShardStatus.CLOSED, IndexShardStatus.OPEN));
+        final FindIndexShardCriteria criteria = FindIndexShardCriteria
+                .builder()
+                .indexShardStatusSet(statusSelection)
+                .build();
+
+        final IndexDocument document1 = new IndexDocument();
+        document1.add(new FieldValue(LuceneIndexField
+                .builder()
+                .fldName("SourcePort")
+                .fldType(FieldType.TEXT)
+                .analyzerType(AnalyzerType.ALPHA_NUMERIC)
+                .termPositions(false)
+                .indexed(true)
+                .stored(false)
+                .build(), ValString.create("12345")));
+        indexer.addDocument(indexShardKey1, document1);
+        indexer.addDocument(indexShardKey1, document1);
+        assertThat(indexShardService.find(criteria).size()).isOne();
+
+        final IndexDocument document2 = new IndexDocument();
+        document2.add(new FieldValue(LuceneIndexField
+                .builder()
+                .fldName("SourcePort")
+                .fldType(FieldType.INTEGER)
+                .analyzerType(AnalyzerType.KEYWORD)
+                .termPositions(false)
+                .indexed(true)
+                .stored(false)
+                .build(), ValInteger.create(12345)));
+        indexer.addDocument(indexShardKey1, document2);
+        indexer.addDocument(indexShardKey1, document2);
+        assertThat(indexShardService.find(criteria).size()).isEqualTo(2);
+
+        final IndexDocument document3 = new IndexDocument();
+        document3.add(new FieldValue(LuceneIndexField
+                .builder()
+                .fldName("SourcePort")
+                .fldType(FieldType.TEXT)
+                .analyzerType(AnalyzerType.KEYWORD)
+                .termPositions(true)
+                .indexed(true)
+                .stored(false)
+                .build(), ValString.create("12345")));
+        indexer.addDocument(indexShardKey1, document3);
+        indexer.addDocument(indexShardKey1, document3);
+        assertThat(indexShardService.find(criteria).size()).isEqualTo(3);
+    }
+
+    @Disabled
+    @Test
+    void testPerformance() {
+        assertThat(indexShardService.find(FindIndexShardCriteria.matchAll()).size()).isZero();
+
+        final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010",
+                commonTestScenarioCreator.createIndexFields(),
+                1000000);
+        final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
+        final IndexShardKey indexShardKey1 = IndexShardKey.createKey(index1);
+
+        final IndexDocument document1 = new IndexDocument();
+        document1.add(new FieldValue(LuceneIndexField
+                .builder()
+                .fldName("SourcePort")
+                .fldType(FieldType.TEXT)
+                .analyzerType(AnalyzerType.ALPHA_NUMERIC)
+                .termPositions(false)
+                .indexed(true)
+                .stored(false)
+                .build(), ValString.create("12345")));
+        for (int i = 0; i < 1000000; i++) {
+            indexer.addDocument(indexShardKey1, document1);
+        }
+    }
+
     private void checkDocCount(final int expected, final IndexShardWriter indexShardWriter) {
         assertThat(indexShardWriter.getDocumentCount()).isEqualTo(expected);
     }
@@ -200,51 +382,5 @@ class TestIndexShardWriterImpl extends AbstractCoreIntegrationTest {
     private boolean compareStatus(final IndexShardStatus expected, final long indexShardId) {
         final IndexShard loaded = indexShardService.loadById(indexShardId);
         return expected.equals(loaded.getStatus());
-    }
-
-    @Test
-    void testSimpleRoll() throws IOException {
-        // Do some work.
-        final IndexDocument document = new IndexDocument();
-        document.add(new FieldValue(LuceneIndexField.createField("test"), ValString.create("test")));
-
-        final DocRef indexRef1 = commonTestScenarioCreator.createIndex("TEST_2010",
-                commonTestScenarioCreator.createIndexFields(),
-                10);
-        final LuceneIndexDoc index1 = indexStore.readDocument(indexRef1);
-        final IndexShardKey indexShardKey1 = IndexShardKeyUtil.createTestKey(index1);
-
-        final IndexShardWriter writer1 = indexShardWriterCache.getWriterByShardKey(indexShardKey1);
-
-        for (int i = 0; i < 10; i++) {
-            assertThat(indexShardWriterCache.getWriterByShardKey(indexShardKey1)).isEqualTo(writer1);
-            indexer.addDocument(indexShardKey1, document);
-        }
-
-        // Make sure the writer is full.
-        assertThatThrownBy(() -> writer1.addDocument(document)).isInstanceOf(IndexException.class);
-
-        // Make sure the writer is still open.
-        assertThat(compareStatus(IndexShardStatus.OPEN, writer1.getIndexShardId())).isTrue();
-
-        // Now push the writer over the edge so we get a new writer.
-        indexer.addDocument(indexShardKey1, document);
-
-        // Get the new writer.
-        final IndexShardWriter writer2 = indexShardWriterCache.getWriterByShardKey(indexShardKey1);
-
-        // Make sure the writers are not the same.
-        assertThat(writer2).isNotEqualTo(writer1);
-
-        for (int i = 1; i < 10; i++) {
-            assertThat(indexShardWriterCache.getWriterByShardKey(indexShardKey1)).isEqualTo(writer2);
-            indexer.addDocument(indexShardKey1, document);
-        }
-
-        // Make sure the writer is full.
-        assertThatThrownBy(() -> writer2.addDocument(document)).isInstanceOf(IndexException.class);
-
-        // Make sure the writer is still open.
-        assertThat(compareStatus(IndexShardStatus.OPEN, writer2.getIndexShardId())).isTrue();
     }
 }