+++ stroom-core-client/src/main/java/stroom/dashboard/client/input/gin/InputModule.java
@@ -22,6 +22,9 @@ import stroom.dashboard.client.input.BasicKeyValueInputSettingsViewImpl;
 import stroom.dashboard.client.input.BasicListInputSettingsPresenter;
 import stroom.dashboard.client.input.BasicListInputSettingsPresenter.BasicListInputSettingsView;
 import stroom.dashboard.client.input.BasicListInputSettingsViewImpl;
+import stroom.dashboard.client.input.BasicTextInputSettingsPresenter;
+import stroom.dashboard.client.input.BasicTextInputSettingsPresenter.BasicTextInputSettingsView;
+import stroom.dashboard.client.input.BasicTextInputSettingsViewImpl;
 import stroom.dashboard.client.input.KeyValueInputPlugin;
 import stroom.dashboard.client.input.KeyValueInputPresenter;
 import stroom.dashboard.client.input.KeyValueInputPresenter.KeyValueInputView;
@@ -30,6 +33,10 @@ import stroom.dashboard.client.input.ListInputPlugin;
 import stroom.dashboard.client.input.ListInputPresenter;
 import stroom.dashboard.client.input.ListInputPresenter.ListInputView;
 import stroom.dashboard.client.input.ListInputViewImpl;
+import stroom.dashboard.client.input.TextInputPlugin;
+import stroom.dashboard.client.input.TextInputPresenter;
+import stroom.dashboard.client.input.TextInputPresenter.TextInputView;
+import stroom.dashboard.client.input.TextInputViewImpl;
 
 import com.gwtplatform.mvp.client.gin.AbstractPresenterModule;
 
@@ -39,16 +46,20 @@ public class InputModule extends AbstractPresenterModule {
     protected void configure() {
         bind(KeyValueInputPlugin.class).asEagerSingleton();
         bindPresenterWidget(KeyValueInputPresenter.class, KeyValueInputView.class, KeyValueInputViewImpl.class);
+        bindPresenterWidget(BasicKeyValueInputSettingsPresenter.class,
+                BasicKeyValueInputSettingsView.class,
+                BasicKeyValueInputSettingsViewImpl.class);
 
         bind(ListInputPlugin.class).asEagerSingleton();
         bindPresenterWidget(ListInputPresenter.class, ListInputView.class, ListInputViewImpl.class);
-
         bindPresenterWidget(BasicListInputSettingsPresenter.class,
                 BasicListInputSettingsView.class,
                 BasicListInputSettingsViewImpl.class);
 
-        bindPresenterWidget(BasicKeyValueInputSettingsPresenter.class,
-                BasicKeyValueInputSettingsView.class,
-                BasicKeyValueInputSettingsViewImpl.class);
+        bind(TextInputPlugin.class).asEagerSingleton();
+        bindPresenterWidget(TextInputPresenter.class, TextInputView.class, TextInputViewImpl.class);
+        bindPresenterWidget(BasicTextInputSettingsPresenter.class,
+                BasicTextInputSettingsView.class,
+                BasicTextInputSettingsViewImpl.class);
     }
 }