+++ stroom-core-shared/src/main/java/stroom/dashboard/shared/ComponentSettings.java
@@ -31,13 +31,16 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
         @JsonSubTypes.Type(value = VisComponentSettings.class, name = "vis"),
         @JsonSubTypes.Type(value = TextComponentSettings.class, name = "text"),
         @JsonSubTypes.Type(value = KeyValueInputComponentSettings.class, name = "key-value-input"),
-        @JsonSubTypes.Type(value = ListInputComponentSettings.class, name = "list-input")
+        @JsonSubTypes.Type(value = ListInputComponentSettings.class, name = "list-input"),
+        @JsonSubTypes.Type(value = TextInputComponentSettings.class, name = "text-input")
 })
 @JsonInclude(Include.NON_NULL)
 public interface ComponentSettings {
+
     Builder copy();
 
     interface Builder {
+
         ComponentSettings build();
     }
 }